Move pixel_functions to util subdir
